requirements.txt
dj-database-url==0.5.0
Django==2.2.10
gunicorn==20.0.4
Pillow==7.0.0
psycopg2-binary==2.8.4
python-slugify==4.0.0
pytz==2019.3
sqlparse==0.3.1
text-unidecode==1.3
whitenoise==5.0.1
